+++ src/solver.cpp
@@ -1773,6 +1773,8 @@ int64_t Solver::get_statistic_value (const char *opt) const {
     return internal->stats.conflicts;
   if (!strcmp (opt, "decisions"))
     return internal->stats.decisions;
+  if (!strcmp (opt, "ticks"))
+    return internal->stats.ticks.search[0] + internal->stats.ticks.search[1];
   if (!strcmp (opt, "propagations"))
     return internal->stats.propagations.search;
   if (!strcmp (opt, "clauses"))